+++ ha-docs/pages/calc_api/token/index.vue
@@ -1,0 +1,90 @@
+<template>
+  <div>
+    <v-row justify="center">
+      <v-col sm="12">
+        <AppBreadCrumbs :items="breadcrumbs" />
+      </v-col>
+    </v-row>
+    <v-row justify="center">
+      <v-col sm="12">
+        <AppContentsTitle :title="breadcrumbs[breadcrumbs.length - 1].text" />
+      </v-col>
+    </v-row>
+    <v-row justify="center">
+      <v-col sm="12">
+        <AppDocs :flow="flow" />
+      </v-col>
+    </v-row>
+  </div>
+</template>
+
+<script>
+import AppBreadCrumbs from "~/components/AppBreadCrumbs.vue";
+import AppContentsTitle from "~/components/AppContentsTitle.vue";
+import AppDocs from "~/components/AppDocs.vue";
+
+export default {
+  // calc_apiのレイアウトを適用
+  layout: "calcApiLayout",
+  components: {
+    AppBreadCrumbs,
+    AppContentsTitle,
+    AppDocs,
+  },
+  data: function () {
+    return {
+      breadcrumbs: [
+        {
+          text: "Top",
+          disabled: false,
+          href: "/",
+        },
+        {
+          text: "calc_api",
+          disabled: false,
+          href: "/calc_api",
+        },
+        {
+          text: "トークン発行API",
+          disabled: true,
+          href: "/calc_api/token",
+        },
+      ],
+      flow: [
+        {
+          id: "1",
+          text: "リクエストログ出力",
+          edgeType: "round",
+          next: ["2"],
+        },
+        {
+          id: "2",
+          text: "トークンを生成",
+          edgeType: "round",
+          next: ["3"],
+        },
+        {
+          id: "3",
+          text: "レスポンスJSON生成",
+          edgeType: "round",
+          next: ["4"],
+        },
+        {
+          id: "4",
+          text: "レスポンスログ出力",
+          edgeType: "round",
+          next: ["5"],
+        },
+        {
+          id: "5",
+          text: "レスポンスJSON返却",
+          edgeType: "round",
+        },
+      ],
+    };
+  },
+};
+</script>
+
+<style scoped>
+</style>
