+++ ha-docs/pages/healthinfoapp/dashboard/breath/index.vue
@@ -87,7 +87,7 @@ export default {
           id: "b4",
           text: "Token発行API 実施",
           edgeType: "round",
-          url: "/node/api/token",
+          url: "/calc_api/token",
           next: ["b5"],
           style: "fill:#ffce9e",
         },
@@ -110,7 +110,7 @@ export default {
           id: "b7",
           text: "肺活量計算API 実施",
           edgeType: "round",
-          url: "/node/api/breath_capacity",
+          url: "/calc_api/breath_capacity",
           next: ["b8"],
           style: "fill:#ffce9e",
         },
