+++ ha-docs/pages/healthinfoapp/dashboard/calorie/index.vue
@@ -87,7 +87,7 @@ export default {
           id: "b4",
           text: "Token発行API 実施",
           edgeType: "round",
-          url: "/node/api/token",
+          url: "/calc_api/token",
           next: ["b5"],
           style: "fill:#ffce9e",
         },
@@ -110,7 +110,7 @@ export default {
           id: "b7",
           text: "カロリー計算API 実施",
           edgeType: "round",
-          url: "/node/api/calorie",
+          url: "/calc_api/calorie",
           next: ["b8"],
           style: "fill:#ffce9e",
         },
